+++ tridydb-cli/src/include/Compressor.js
@@ -5,7 +5,7 @@ import * as common from '../utility/common.js';
 export class Compressor {
     constructor() { }
 
-    static _compressModuleLite(target, opts = { }) {
+    static _compressModuleLite(target, alias, opts = { }) {
         opts.typeless = opts.typeless ?? false;
         opts.tagless  = opts.tagless  ?? false;
 
@@ -14,25 +14,18 @@ export class Compressor {
         }
 
         if (opts.typeless) {
-            target.enterSetAndLeave(common.global.alias.type, undefined);
+            target.enterDeleteAndLeave(alias.type);
         }
         if (opts.tagless) {
-            target.enterSetAndLeave(common.global.alias.tags, undefined);
+            target.enterDeleteAndLeave(alias.tags);
         }
 
-        target.enterPos(common.global.alias.nested);
-        if (!target.isPosEmpty()) {
-            target.enterPos(0);
-            while (!target.isPosUndefined()) {
-                this._compressModuleLite(target, opts);
-                target.nextItem();
-            }
-            target.leavePos();
-        }
-        target.leavePos();
+        target.traverse(() => {
+            this._compressModuleLite(target, alias, opts);
+        });
     }
 
-    static _compressModuleHeavy(target, opts = { }) {
+    static _compressModuleHeavy(target, alias, opts = { }) {
         opts.strict = opts.strict ?? false;
 
         if (!common.isDictionary(target)) {
@@ -49,7 +42,7 @@ export class Compressor {
          * That is even if we "could" reduce these down to simply "this" and "that".
          * The free data structure is meant to contain arbitrary user data that should be altered by TridyDB as little as possible.
          */
-        let free                = target[common.global.alias.state];
+        let free                = target[alias.state];
         let reduce_to_array     = true;
         let reduce_to_primitive = true;
         let reduce_to_nothing   = true;
@@ -66,38 +59,42 @@ export class Compressor {
         }
         free = common.toDictionary(free);
 
-        const tree = common.toArray(target[common.global.alias.nested]);
+        const tree = common.toArray(target[alias.nested]);
 
         target = new StateTree(free);
 
-        for (let sub of tree) {
+        for (let mod of tree) {
             // Don't move "sub =" before "type ="; "_compressModuleHeavy()" truncates the type specifier.
-            type = common.isDictionary(sub) ? sub[common.global.alias.type] : null;
+            type = common.isDictionary(mod) ? mod[alias.type] : null;
 
-            sub = this._compressModuleHeavy(sub, opts);
-            if (sub === undefined) {
+            mod = this._compressModuleHeavy(mod, alias, opts);
+            if (mod === undefined) {
                 continue;
+            }
+
+            if (target.enterGetAndLeave('_xml') === true) {
+                target.enterPos('unparsed');
             }
             
             if (common.isPrimitive(type)) {
                 target.enterPos(type);
                 if (opts.strict) {
                     if (common.isArray(target.getPosValue()) || target.isPosUndefined()) {
-                        target.putPosValue(sub);
+                        target.putPosValue(mod);
                     } else {
                         target.leavePos();
                         target.enterPos(0);
                         while (!target.isPosUndefined()) {
                             target.nextItem();
                         }
-                        target.setPosValue(sub);
+                        target.setPosValue(mod);
                         target.leavePos();
                         target.enterPos(type);
                     }
                 } else if (target.isPosUndefined()) {
-                    target.setPosValue(sub);
+                    target.setPosValue(mod);
                 } else {
-                    target.putPosValue(sub);
+                    target.putPosValue(mod);
                 }
                 target.leavePos();
             } else {
@@ -106,10 +103,14 @@ export class Compressor {
                     target.nextItem();
                 }
                 if (opts.strict) {
-                    target.putPosValue(sub);
+                    target.putPosValue(mod);
                 } else {
-                    target.setPosValue(sub);
+                    target.setPosValue(mod);
                 }
+                target.leavePos();
+            }
+
+            if (target.enterGetAndLeave('_xml') === true) {
                 target.leavePos();
             }
         }
@@ -135,30 +136,30 @@ export class Compressor {
         return target;
     }
 
-    static compressModule(module, lvl = 0) {
+    static compressModule(module, alias, lvl = 0) {
         module = common.deepCopy(module);
 
         switch (lvl) {
             case 1:
-                module = new StateTree(module);
-                this._compressModuleLite(module, { typeless: true, tagless: false });
+                module = new StateTree(module, alias);
+                this._compressModuleLite(module, alias, { typeless: true, tagless: false });
                 module = module.getRaw();
                 break;
             case 2:
-                module = new StateTree(module);
-                this._compressModuleLite(module, { typeless: false, tagless: true });
+                module = new StateTree(module, alias);
+                this._compressModuleLite(module, alias, { typeless: false, tagless: true });
                 module = module.getRaw();
                 break;
             case 3:
-                module = new StateTree(module);
-                this._compressModuleLite(module, { typeless: true, tagless: true });
+                module = new StateTree(module, alias);
+                this._compressModuleLite(module, alias, { typeless: true, tagless: true });
                 module = module.getRaw();
                 break;
             case 4:
-                module = this._compressModuleHeavy(module, { strict: true });
+                module = this._compressModuleHeavy(module, alias, { strict: true });
                 break;
             case 5:
-                module = this._compressModuleHeavy(module, { strict: false });
+                module = this._compressModuleHeavy(module, alias, { strict: false });
                 break;
         }